+++ src/parser/enrichers/trait/goliath/FrostsChill.mjs
@@ -10,6 +10,7 @@ export default class FrostsChill extends DDBEnricherData {
     return {
       targetType: "creature",
       activationType: "special",
+      addItemConsume: true,
       data: {
         damage: {
           parts: [
@@ -32,5 +33,18 @@ export default class FrostsChill extends DDBEnricherData {
         ],
       },
     ];
+  }
+
+  get override() {
+    return {
+      data: {
+        "system.uses": this._getUsesWithSpent({
+          type: "race",
+          name: this.ddbParser.originalName,
+          max: "@prof",
+          period: "lr",
+        }),
+      },
+    };
   }
 }